refactor player fsm
