+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm.cs
@@ -10,39 +10,8 @@ using UnityEngine.PlayerLoop;
 using UnityEngine.Serialization;
 using Wasp;
 
-public class PlayerFsm : GravityFsm
+public partial class PlayerFsm : GravityFsm
 {
-    
-    protected override void OnStart()
-    {
-        print("player onstart");
-        base.OnStart();
-        Singleton = this;
-        InitState = PlayerFsmState.GroundMove;
-        print("init state: " + InitState);
-        _movementAnimationMirror = false;
-        TryGetComponent(out _playerInput);
-        _inputBuffer = new InputBuffer(_playerInput, 0.275f);
-        _inputBuffer.InitInput("Jump");
-        _inputBuffer.InitInput("Dash");
-        _inputBuffer.InitInput("Attack");
-        _camera = Camera.main;
-        _previousWallrunSide = FlankType.None;
-        
-        // QualitySettings.vSyncCount = 0; // Set vSyncCount to 0 so that using .targetFrameRate is enabled.
-        // Application.targetFrameRate = 30;
-        
-    }
-    
-    protected override void OnAwake()
-    {
-        Singleton = this;
-    }
-    
-    // --------- End of boilerplate Fsm code --------- //
-    
-    // --------- Subclass Fsm data ------------- //
-    
     public class PlayerFsmState : GravityFsmState
     {
         public static int GroundMove;
@@ -68,6 +37,7 @@ public class PlayerFsm : GravityFsm
         public static int GrappleStartup;
         public static int GrappleFlipsquat;
         public static int GrappleFlip;
+        public static int VaultHang;
     }
 
     public class PlayerFsmTrigger : GravityFsmTrigger
@@ -86,655 +56,37 @@ public class PlayerFsm : GravityFsm
         public static int Attack;
         public static int ContactHitboxTrigger;
     }
-
-    private enum FlankType
+    
+    protected override void OnAwake()
     {
-        Left,
-        Right,
-        None
+        Singleton = this;
     }
     
-    // Fluid data
-    
-    private PlayerInput _playerInput;
-    private InputBuffer _inputBuffer;
-    private Camera _camera;
-    private float _momentum = 0f;
-    private float _stateEntryMomentum = 0f;
-    private Vector3 _currentLedgePosition;
-    private Vector3 _currentFlankWallNormal;
-    private FlankType _currentFlankType;
-    private FlankType _previousWallrunSide;
-    private Vector3 _checkpointVector3;
-    private Quaternion _checkpointQuaternion;
-    private bool _movementAnimationMirror;
-    public static PlayerFsm Singleton;
-    
-    // Events
-    
-    public static event Action<float> OnPlayerMomentumUpdated;
-    public static event Action<Vector3, bool> OnPlayerPositionUpdated;
-    public static event Action OnPlayerImpaleStateEntered;
-    
-    // Input
-    
-    public const float InputMagnitudeThreshhold = 0.1f;
-    
-    // Raycasting
-    
-    private const float ForwardRaycastDistance = 0.9f;
-    private const float DynamicForwardRaycastMaximumModifier = 2f;
-    private const float CollisionMoveSphereCastRadius = 0.5f;
-    private const float GroundCollisionMoveSphereCastHeight = 0.95f;
-    private const float FallingCollisionMoveSphereCastHeight = -0.5f;
-    private const float FallingCollisionMoveSphereCastHeightYVelocityThreshhold = -10f;
-    private const float CollisionMoveSphereCastDistance = 0.45f;
-    private const float FaceLedgeHeight = 0.2f;
-    private const float FaceHighLedgeHeight = 2.15f;
-    private const float FaceWallHeight = 2.4f;
-    private const float FaceWallMaximumAngle = 60f;
-    private const float FaceWallStrictMaximumAngle = 20f;
-    private const float MaximumFlankWallDistance = 7.5f;
-    private const float FlankWallHeight = 3f;
-    private const float FlankWallOpenYOffset = -2f;
-    private const float FlankMaximumAngle = 40f;
-    private const float ForceWallRotationRaycastDistance = 3f;
-    
-    // General movement
-    
-    public const float MaxMomentum = 15f;
-    private const float MoveSpeed = 5f;
-    private const float MaximumMomentumSpeedMod = 3.5f;
-    private const float RotationSpeed = 3f;
-    private const float CollisionMomentumLossRate = 300f;
-    private const float MomentumGainRate = 9f;
-    private const float MomentumLossRate = 20f; 
-    private const float MomentumTurnLoss = 5f; 
-    private const float NoMomentumThreshold = 0.25f;
-    private const float LowMomentumThreshhold = 4.75f;
-    private const float LowMomentumRotationMod = 3f;
-    private const float LowMomentumMomentumGainMod = 1.15f;
-    private const float LowMomentumMomentumLossMod = 1.25f;
-    private const float GroundMoveMaximumAnimatorSpeedMod = 3.5f;
-    
-    // Air movement
-    
-    private const float JumpYVelocity = 22f;
-    private const float CoyoteTime = 0.04f;
-    
-    // Wall interaction
-    
-    private const float UpdateLedgePositionEpsilon = 3f;
-    private const float VaultMinimumYVelocity = -2f;
-    private const float VaultMinimumMomentum = 6f;
-    private const float VaultHangLedgeYOffset = -2.5f;
-    private const float VaultHangLedgeLerpStrength = 60f;
-    private const float VaultTurningMultiplier = 0.75f;
-    private const float VaultMinimumAnimatorSpeedMod = 0.3f;
-    private const float VaultMaximumAnimatorSpeedMod = 1.1f;
-    private const float VaultLedgeLerpStrength = 40f;
-    private const float MediumVaultHangMinimumYVelocity = 12f;
-    private const float SlowVaultFinishLedgeLerpStrength = 25f;
-    private const float SlowVaultFinishForwardSpeed = 3f;
-    private const float WallsquatMinimumYVelocity = 10f; 
-    private const float WallSquatMinimumMomentum = 3f;
-    private const float WallstepMinimumYVelocityGain = 12f;
-    private const float WallstepMaximumYVelocityGain = 23.5f;
-    private const float WallstepMinimumDuration = 0.25f;
-    private const float ForceWallRotationSpeed = 3f;
-    private const float WallRunMinimumEntryMomentum = 9f;
-    private const float WallRunMinimumMomentum = 9f;
-    private const float WallRunMinimumYVelocity = 13f; // It's pretty important that this value is larger than WallsquatMinimumYVelocity
-    private const float FlankAlignmentRotationSpeed = 25f;
-    private const float FlankWallVacuumStrength = 20f;
-    private const float WallrunJumpAngle = 75f;
-    
-    // Hard land
-    
-    private const float HardLandAirDiff = -9;
-    private const float HardLandExitMomentum = 4f;
-    private const float HardLandRollExitMomentum = 10f;
-    private const float HardLandRollMinimumMomentum = 7f;
-    private const float HardLandRollForwardSpeed = 14f;
-
-    
-    // Hard turn
-    
-    private const float HardTurnMinimumAngle = 130f;
-    private const float HardTurnMinimumMomentum = 8.5f;
-    private const float HardTurnMomentumLossModifier = 1.25f;
-    
-    // Dash
-
-    private const float DashEntryMomentumGain = 5f;
-    private const float DashEntryMinimumMomentum = 12f;
-    private const float DashsquatTurnMultiplier = 2.25f;
-    private const float DashForwardSpeed = 20f;
-    
-    // Attack
-
-    private const float ImpaleMovementModifier = 1f;
-    private const float ImpaleMomentumOffset = 2.5f;
-    private const float ImpaleMinimumMomentumAfterOffset = 6f;
-    private const float ImpaleMomentumLerpStrenth = 10f;
-    private const float GrappleStartupRotationLerpStrength = 13f;
-    private const float GrappleStartupYPositionLerpStrength = 3f;
-    private const float GrappleStartupYPositionOffset = 1f;
-    private const float GrappleStartupMomentumLossMod = 1.25f;
-    
-    // --------- End of subclass Fsm data ------------- //
-
-
-    public override void SetupMachine()
+    protected override void OnStart()
     {
-        print("player setupmachine");
-        base.SetupMachine();
+        print("player onstart");
+        base.OnStart();
+        Singleton = this;
+        InitState = PlayerFsmState.GroundMove;
+        print("init state: " + InitState);
+        _movementAnimationMirror = false;
+        TryGetComponent(out _playerInput);
+        _inputBuffer = new InputBuffer(_playerInput, 0.275f);
+        _inputBuffer.InitInput("Jump");
+        _inputBuffer.InitInput("Dash");
+        _inputBuffer.InitInput("Attack");
+        _camera = Camera.main;
+        _previousWallrunSide = FlankType.None;
         
-        Machine.Configure(PlayerFsmState.GroundMove)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(PlayerFsmTrigger.HardTurn, PlayerFsmState.HardTurn)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.ImpaleGround, CanImpale)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("GroundMove");
-            });
+        // QualitySettings.vSyncCount = 0; // Set vSyncCount to 0 so that using .targetFrameRate is enabled.
+        // Application.targetFrameRate = 30;
         
-        Machine.Configure(PlayerFsmState.Jumpsquat)
-            .SubstateOf(GravityFsmState.Grounded)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => true)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Jump)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("Jumpsquat");
-                _inputBuffer.ConsumeBuffer("Jump");
-            })
-            .OnExitFrom(FsmTrigger.Timeout,_ =>
-            {
-                YVelocity = JumpYVelocity;
-            });
-        
-        Machine.Configure(PlayerFsmState.Landsquat)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("Landsquat");
-            })
-            .OnExit(_ =>
-            {
-                _movementAnimationMirror = !_movementAnimationMirror;
-                _previousWallrunSide = FlankType.None;
-                _currentFlankType = FlankType.None;
-                var flip = _movementAnimationMirror ? 0 : 1f;
-                Animator.SetFloat("Flip", flip);
-            });
-        
-        Machine.Configure(PlayerFsmState.HardLand)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                _momentum = HardLandExitMomentum;
-                ReplaceAnimatorTrigger("HardLand");
-            });
-        
-        Machine.Configure(PlayerFsmState.HardLandRoll)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                _momentum = HardLandRollExitMomentum;
-                ReplaceAnimatorTrigger("HardLandRoll");
-            });
-
-        
-        Machine.Configure(PlayerFsmState.Jump)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff, 1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll, _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FlankWall, PlayerFsmState.Wallrun, _ => _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.ImpaleAir, CanImpale)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-        // _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity
-            .Permit(PlayerFsmTrigger.Dash, PlayerFsmState.Dashsquat)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("Jump");
-            });
-        
-        
-        Machine.Configure(PlayerFsmState.Fall)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat, _ => TimeInAir <= CoyoteTime)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff, 1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll, _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .Permit(PlayerFsmTrigger.Dash, PlayerFsmState.Dashsquat)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("Fall");
-            });
-        
-        Machine.Configure(PlayerFsmState.HardTurn)
-            .Permit(PlayerFsmTrigger.NoMomentum, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .SubstateOf(GravityFsmState.Grounded)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("HardTurn");
-            });
-
-        Machine.Configure(GravityFsmState.Aerial)
-            ;
-
-        Machine.Configure(PlayerFsmState.Vault)
-            .SubstateOf(GravityFsmState.Aerial)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            .OnEntry(_ =>
-            {
-                UpdateLedgePosition(FaceLedgeHeight);
-                ReplaceAnimatorTrigger("Vault");
-                YVelocity = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.SlowVaultHang)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.SlowVaultFinish)
-            .OnEntry(_ =>
-            {
-                UpdateLedgePosition(FaceHighLedgeHeight);
-                ReplaceAnimatorTrigger("SlowVaultHang");
-                YVelocity = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.MediumVaultHang)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.SlowVaultFinish)
-            .OnEntry(_ =>
-            {
-                if (!UpdateLedgePosition(FaceHighLedgeHeight)) UpdateLedgePosition(FaceLedgeHeight);
-                ReplaceAnimatorTrigger("MediumVaultHang");
-                YVelocity = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.SlowVaultFinish)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            // .SubstateOf(PlayerFsmState.Grounded)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("SlowVaultFinish");
-                YVelocity = 0;
-            })
-            .OnExit(_ =>
-            {
-                _momentum = 3f;
-            });
-        
-        Machine.Configure(PlayerFsmState.Wallstep)
-            .SubstateOf(GravityFsmState.Aerial)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .Permit(GravityFsmTrigger.StartFrameWithNegativeYVelocity, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.SlowVaultHang, _ => YVelocity < MediumVaultHangMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.MediumVaultHang, _ => YVelocity > MediumVaultHangMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => YVelocity > MediumVaultHangMinimumYVelocity, 1)
-            .OnEntry(_ =>
-            {
-                _inputBuffer.ConsumeBuffer("Jump");
-                ReplaceAnimatorTrigger("Wallstep");
-                YVelocity = Mathf.Lerp(WallstepMinimumYVelocityGain, WallstepMaximumYVelocityGain, ComputeMomentumWeight());
-                Animator.SetFloat("VerticalMomentum", ComputeMomentumWeight());
-                _momentum = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.Wallsquat)
-            .SubstateOf(GravityFsmState.Aerial)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            // .Permit(PlayerFsmTrigger.FaceOpen, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.Jump, PlayerFsmState.Wallstep, _ => TimeInCurrentState() > WallstepMinimumDuration, 1)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            .OnEntry(_ =>
-            {
-                YVelocity = 0;
-                ReplaceAnimatorTrigger("Wallsquat");
-            })
-            .OnExitFrom(PlayerFsmTrigger.FaceOpen, _ =>
-            {
-                _momentum = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.Wallrun)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(PlayerFsmTrigger.FlankOpen, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum)
-            .OnEntry(_ =>
-            {
-                print("Yvelocity on entering wallrun: " + YVelocity);
-                _momentum = Mathf.Max(_momentum, WallRunMinimumEntryMomentum);
-                ReplaceAnimatorTrigger("Wallrun");
-
-            })
-            .OnExitFrom(PlayerFsmTrigger.Jump, _ =>
-            {
-                var rotationMod = _currentFlankType == FlankType.Left ? -1f : 1f;
-                var forward = Quaternion.Euler(0f, WallrunJumpAngle * rotationMod, 0f) * _currentFlankWallNormal;
-                transform.rotation = Quaternion.LookRotation(forward, Vector3.up);
-            })
-            .OnExit(_ =>
-            {
-                _previousWallrunSide = _currentFlankType;
-            });
-        
-        Machine.Configure(PlayerFsmState.Dashsquat)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Grapple)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .OnEntry(_ =>
-            {
-                _inputBuffer.ConsumeBuffer("Dash");
-                ReplaceAnimatorTrigger("Dashsquat");
-            });
-        
-        Machine.Configure(PlayerFsmState.Grapple)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GrappleFlipsquat)
-            // .Permit(PlayerFsmTrigger.ContactHitboxTrigger, PlayerFsmState.GrappleFlipsquat)
-            // .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => true)
-            // .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat, _ => true)
-            // .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => true)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .OnEntry(_ =>
-            {
-                YVelocity = 0;
-                _momentum = Mathf.Min(Mathf.Max(_momentum + DashEntryMomentumGain, DashEntryMinimumMomentum), MaxMomentum);
-                transform.rotation =
-                    Quaternion.LookRotation(PlayerWeaponFsm.Singleton.transform.position - transform.position,
-                        Vector3.up);
-                ReplaceAnimatorTrigger("Dash");
-            })
-            .OnExit(_ =>
-            {
-                transform.position = PlayerWeaponFsm.Singleton.transform.position - transform.forward * 0.75f;
-            });
-        
-
-        Machine.Configure(PlayerFsmState.ImpaleGround)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-            .OnEntry(_ =>
-            {
-                Animator.SetLayerWeight(1, 0);
-                _inputBuffer.ConsumeBuffer("Attack");
-                OnPlayerImpaleStateEntered?.Invoke();
-                Animator.SetTrigger("Impale");
-                _stateEntryMomentum = _momentum;
-            }).OnExit(_ =>
-            {
-                Animator.ResetTrigger("Impale");
-            });;
-
-        Machine.Configure(PlayerFsmState.ImpaleAir)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff, 1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll, _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple)
-            .OnEntry(_ =>
-            {
-                Animator.SetLayerWeight(1, 0);
-                _inputBuffer.ConsumeBuffer("Attack");
-                OnPlayerImpaleStateEntered?.Invoke();
-                Animator.SetTrigger("ImpaleJump");
-                _stateEntryMomentum = _momentum;
-            }).OnExit(_ =>
-            {
-                Animator.ResetTrigger("ImpaleJump");
-            });;
-        
-        Machine.Configure(PlayerFsmState.GrappleStartup)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Grapple)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("GrappleStartup");
-                YVelocity = 10;
-                _inputBuffer.ConsumeBuffer("Attack");
-            }).OnExit(_ =>
-            {
-                YVelocity = 0;
-            });
-        
-        Machine.Configure(PlayerFsmState.GrappleFlipsquat)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GrappleFlip)
-            .OnEntry(_ =>
-            {
-                
-                // transform.DOShakePosition(0.5f, 0.3f);
-                ReplaceAnimatorTrigger("GrappleFlipsquat");
-                // HitstopManager.Singleton.StartHitstop(0.075f);
-            })
-            .OnExit(_ =>
-            {
-                
-            });
-        
-        Machine.Configure(PlayerFsmState.GrappleFlip)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff, 1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll, _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat, _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FlankWall, PlayerFsmState.Wallrun, _ => _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity)
-            .OnEntry(_ =>
-            {
-                _momentum = 10f;
-                ReplaceAnimatorTrigger("GrappleFlip");
-                YVelocity = 30;
-            });
-    }
-
-    public override void SetupStateMaps()
-    {
-        print("player setupstatemaps");
-        
-        base.SetupStateMaps();
-        StateMapConfig.Duration.Add(PlayerFsmState.Jumpsquat, 0.175f);
-        StateMapConfig.Duration.Add(PlayerFsmState.Landsquat, 0.125f);
-        StateMapConfig.Duration.Add(PlayerFsmState.HardLand, 0.65f);
-        StateMapConfig.Duration.Add(PlayerFsmState.HardLandRoll, 0.45f);
-        StateMapConfig.Duration.Add(PlayerFsmState.Vault, 0.25f);
-        StateMapConfig.Duration.Add(PlayerFsmState.SlowVaultHang, 0.975f);
-        StateMapConfig.Duration.Add(PlayerFsmState.MediumVaultHang, 0.375f);
-        StateMapConfig.Duration.Add(PlayerFsmState.SlowVaultFinish, 0.3f);
-        StateMapConfig.Duration.Add(PlayerFsmState.Wallsquat, 0.55f);
-        StateMapConfig.Duration.Add(PlayerFsmState.Dashsquat, 0.1f);
-        StateMapConfig.Duration.Add(PlayerFsmState.Grapple, 0.1f);
-        StateMapConfig.Duration.Add(PlayerFsmState.ImpaleGround, 0.55f);
-        StateMapConfig.Duration.Add(PlayerFsmState.ImpaleAir, 0.55f);
-        StateMapConfig.Duration.Add(PlayerFsmState.GrappleStartup, 0.175f);
-        StateMapConfig.Duration.Add(PlayerFsmState.GrappleFlipsquat, 0.265f);
-        
-        StateMapConfig.GravityStrengthMod.Add(PlayerFsmState.Wallstep, 0.5f);
-        StateMapConfig.GravityStrengthMod.Add(PlayerFsmState.Wallrun, 0.55f);
-    }
-
-    public override void OnFireTriggers()
-    {
-        
-        print("player onfiretriggers");
-        
-        base.OnFireTriggers();
-        
-        if (_inputBuffer.IsBuffered("Jump"))
-        {
-            Machine.Fire(PlayerFsmTrigger.Jump);
-        }
-        
-        if (_inputBuffer.IsBuffered("Attack"))
-        {
-            Machine.Fire(PlayerFsmTrigger.Attack);
-        }
-        
-        // if (_inputBuffer.IsBuffered("Dash"))
-        // {
-        //     Machine.Fire(PlayerFsmTrigger.Dash);
-        // }
-        
-        var v3 = GetInputMovementVector3();
-        var angle = Vector3.Angle(v3.normalized, transform.forward.normalized);
-        if (angle > HardTurnMinimumAngle && _momentum > HardTurnMinimumMomentum)
-        {
-            Machine.Fire(PlayerFsmTrigger.HardTurn);
-        }
-
-        if (_momentum < NoMomentumThreshold)
-        {
-            Machine.Fire(PlayerFsmTrigger.NoMomentum);
-        }
-
-        FireFaceTriggers();
-        // FireFlankTriggers();
-    }
-
-    private void FireFaceTriggers()
-    {
-        var forwardRaycastDistance = ComputeDynamicForwardRaycastDistance();
-        if (Physics.Raycast(transform.position + Vector3.up * FaceWallHeight, transform.forward, 
-                out var hit, forwardRaycastDistance, ~0, QueryTriggerInteraction.Ignore) && Vector3.Angle(-hit.normal, transform.forward) < FaceWallMaximumAngle)
-        {
-            Machine.Fire(Vector3.Angle(-hit.normal, transform.forward) < FaceWallStrictMaximumAngle
-                ? PlayerFsmTrigger.FaceWallStrict
-                : PlayerFsmTrigger.FaceWall);
-        } else if (Physics.Raycast(transform.position + Vector3.up * FaceHighLedgeHeight, transform.forward, 
-                       out hit, forwardRaycastDistance, ~0, QueryTriggerInteraction.Ignore) && Vector3.Angle(-hit.normal, transform.forward) < FaceWallMaximumAngle)
-        {
-            Machine.Fire(PlayerFsmTrigger.FaceHighLedge);
-        } else if (Physics.Raycast(transform.position + Vector3.up * FaceLedgeHeight, transform.forward,
-                       out hit, forwardRaycastDistance, ~0, QueryTriggerInteraction.Ignore) && Vector3.Angle(-hit.normal, transform.forward) < FaceWallMaximumAngle)
-        {
-            var slope = Vector3.Angle(hit.normal, Vector3.up);
-            if (slope > 70f) Machine.Fire(PlayerFsmTrigger.FaceLedge);
-        }
-        else
-        {
-            Machine.Fire(PlayerFsmTrigger.FaceOpen);
-        }
-        
-        // Debug.DrawRay(transform.position + Vector3.up * FaceWallHeight, transform.forward * forwardRaycastDistance, Color.red);
-        // Debug.DrawRay(transform.position + Vector3.up * FaceHighLedgeHeight, transform.forward * forwardRaycastDistance, Color.yellow);
-        // Debug.DrawRay(transform.position + Vector3.up * FaceLedgeHeight, transform.forward * forwardRaycastDistance, Color.cyan);
-    }
-
-    private void FireFlankTriggers()
-    {
-        RaycastHit hit;
-        var maximumFlankRaycastDistance = MaximumFlankWallDistance;
-        var flankRaycastOrigin = transform.position + Vector3.up * FlankWallHeight;
-        if 
-            (Physics.Raycast(flankRaycastOrigin, transform.right,
-                 out hit, maximumFlankRaycastDistance, ~0, QueryTriggerInteraction.Ignore) 
-             && IsHitValidFlank(hit, true) && _previousWallrunSide != FlankType.Right)
-        {
-            Machine.Fire(PlayerFsmTrigger.FlankWall);
-            _currentFlankWallNormal = hit.normal;
-            _currentFlankType = FlankType.Right;
-            Animator.SetFloat("Flip", 0);
-
-        } else if 
-            (Physics.Raycast(flankRaycastOrigin, -transform.right, 
-                 out hit, maximumFlankRaycastDistance, ~0, QueryTriggerInteraction.Ignore)
-             && IsHitValidFlank(hit, false) && _previousWallrunSide != FlankType.Left)
-        {
-            Machine.Fire(PlayerFsmTrigger.FlankWall);
-            _currentFlankWallNormal = hit.normal;
-            _currentFlankType = FlankType.Left;
-            Animator.SetFloat("Flip", 1);
-        }
-        else if (!Physics.Raycast(flankRaycastOrigin + (Vector3.up * FlankWallOpenYOffset), transform.right,
-                     out hit, maximumFlankRaycastDistance, ~0, QueryTriggerInteraction.Ignore))
-        {
-            Machine.Fire(PlayerFsmTrigger.FlankOpen);
-        }
-        
-        Debug.DrawRay(flankRaycastOrigin, transform.right * maximumFlankRaycastDistance, Color.blue);
-        Debug.DrawRay(flankRaycastOrigin, -transform.right * maximumFlankRaycastDistance, Color.blue);
     }
     
-
-    private bool IsHitValidFlank(RaycastHit hit, bool left)
-    {
-        float DistanceFromPointToPlane(Vector3 pointA, Vector3 pointB, Vector3 planeNormal)
-        {
-            Vector3 AB = pointA - pointB;
-            float distance = Vector3.Dot(AB, planeNormal.normalized);
-            return Mathf.Abs(distance); // Absolute value to get the unsigned distance
-        }
-        
-        float flipMod = left ? -1f : 1f;
-        var distance = DistanceFromPointToPlane(transform.position, hit.point, hit.normal);
-        var angle = Vector3.Angle(transform.right * flipMod, hit.normal);
-        print(angle);
-        return distance < 2.5f && angle < FlankMaximumAngle;;
-    }
-
-    private float ComputeDynamicForwardRaycastDistance()
-    {
-        return Mathf.Lerp(1f, DynamicForwardRaycastMaximumModifier, ComputeMomentumWeight()) * ForwardRaycastDistance * GetRaycastTimeModifier();
-    }
-
 
     public override void OnUpdate()
     {
-        
-        if (HitstopManager.Singleton.IsHitstopActive())
-        {
-            Animator.enabled = false;
-            return;
-        }
-        else
-        {
-            Animator.enabled = true;
-        }
-        
-        print("player onupdate");
-        
+        if (HitstopOnUpdate()) return;
         
         base.OnUpdate();
         _inputBuffer.OnUpdate();
@@ -743,7 +95,6 @@ public class PlayerFsm : GravityFsm
                                                             Machine.IsInState(PlayerFsmState.ForceWallRotation) ||
                                                             YVelocity < -6f);
         
-        print(Machine.State());
         if (Machine.IsInState(PlayerFsmState.GroundMove))
         {
             
@@ -894,239 +245,19 @@ public class PlayerFsm : GravityFsm
         }
     }
 
-    private void MoveYOntoLedge(float yOffset, float lerpStrength)
+    private bool HitstopOnUpdate()
     {
-        var newY = lerpStrength < 0 ? _currentLedgePosition.y : Mathf.Lerp(transform.position.y, _currentLedgePosition.y + yOffset, Time.deltaTime * lerpStrength);
-        transform.position = new Vector3(transform.position.x, newY, transform.position.z);
-    }
-
-    private bool UpdateLedgePosition(float ledgeHeight)
-    {
-        var downwardRaycastOrigin = transform.position + (Vector3.up * (ledgeHeight + UpdateLedgePositionEpsilon)) + transform.forward * ComputeDynamicForwardRaycastDistance();
-        Debug.DrawLine(downwardRaycastOrigin, downwardRaycastOrigin - (Vector3.up * (ledgeHeight + UpdateLedgePositionEpsilon)), Color.green);
-
-        if (!Physics.Raycast(downwardRaycastOrigin, -Vector3.up, out var hit,
-                (ledgeHeight + UpdateLedgePositionEpsilon) * GetRaycastTimeModifier(), ~0, QueryTriggerInteraction.Ignore)) return false;
-        _currentLedgePosition = hit.point;
-        return true;
-
-    }
-    
-
-
-    private void HandleFlankAlignment()
-    {
-        if (_currentFlankType == FlankType.None) return; 
-        var rotationMod = _currentFlankType == FlankType.Left ? -1f : 1f;
-        var forward = Quaternion.Euler(0f, 90f * rotationMod, 0f) * _currentFlankWallNormal;
-        Debug.DrawRay(transform.position, forward, Color.yellow);
-        var lookRotation = Quaternion.LookRotation(forward, Vector3.up);
-        transform.rotation = Quaternion.Slerp(transform.rotation, lookRotation, Time.deltaTime * FlankAlignmentRotationSpeed);
-    }
-
-    private void HandleTurning(float multiplier = 1f, bool forceForwardInput = false)
-    {
-        
-        var v3 = GetInputMovementVector3();
-        var inputVector3 = forceForwardInput ? MirrorInputForward(v3, transform.forward) : v3;
-        
-        var v2 = GetInputMovementVector2();
-        if (v2.magnitude < InputMagnitudeThreshhold)
+        if (HitstopManager.Singleton.IsHitstopActive())
         {
-            inputVector3 = transform.forward;
-        }
-        
-        float momentumWeight = ComputeMomentumWeight();
-        var angle = Vector3.SignedAngle(inputVector3.normalized, transform.forward.normalized, Vector3.up);
-        var animationDesiredTurnAmount = Mathf.InverseLerp(35f, -35f, angle);
-        animationDesiredTurnAmount = Mathf.Lerp(-1, 1, animationDesiredTurnAmount);
-        var turnAmount = Animator.GetFloat("TurnAmount");
-        var turnLerpSpeed = Mathf.Abs(animationDesiredTurnAmount) > Mathf.Abs(turnAmount) ? 10f : 2f;
-        Animator.SetFloat("TurnAmount", Mathf.Lerp(turnAmount, animationDesiredTurnAmount, Time.deltaTime * turnLerpSpeed));
-        Animator.SetLayerWeight(1, Mathf.Abs(turnAmount) * momentumWeight);
-            
-        var momentumDesiredTurnAmount = Mathf.InverseLerp(170f, -170f, angle);
-        momentumDesiredTurnAmount = Mathf.Lerp(-1, 1, momentumDesiredTurnAmount);
-        _momentum = Mathf.Max(0, _momentum - (MomentumLossRate * Time.deltaTime *
-                                              Mathf.Abs(momentumDesiredTurnAmount) * momentumWeight * MomentumTurnLoss));
-        
-        var quaternion = Quaternion.LookRotation(inputVector3.normalized, Vector3.up);
-        
-        var lowMomentumRotationMod = _momentum < LowMomentumThreshhold ? LowMomentumRotationMod : 1f;
-        transform.rotation = Quaternion.Slerp(transform.rotation, quaternion, RotationSpeed * Time.deltaTime * lowMomentumRotationMod * multiplier);
-    }
-
-    private void HandleInputMomentumChange()
-    {
-        var v2 = GetInputMovementVector2();
-        if (v2.magnitude > InputMagnitudeThreshhold)
-        {
-            var lowMomentumMomentumGainMod = _momentum < LowMomentumThreshhold ? LowMomentumMomentumGainMod : 1f;
-            _momentum = Mathf.Min(MaxMomentum, _momentum + MomentumGainRate  * lowMomentumMomentumGainMod *  Time.deltaTime);
+            Animator.enabled = false;
+            return true;
         }
         else
         {
-            var lowMomentumMomentumLossMod = _momentum < LowMomentumThreshhold ? LowMomentumMomentumLossMod : 1f;
-            _momentum = Mathf.Max(0, _momentum - (MomentumLossRate * lowMomentumMomentumLossMod * Time.deltaTime));
+            Animator.enabled = true;
         }
-    }
 
-
-    // ------------ Helper functions ------------- //
-    
-    private Vector2 GetInputMovementVector2()
-    {
-        return _playerInput.actions["Move"].ReadValue<Vector2>();
-    }
-    
-    public Vector3 GetInputMovementVector3()
-    {
-        var v2 = GetInputMovementVector2();
-        return Quaternion.Euler(0, _camera.transform.rotation.eulerAngles.y, 0) * new Vector3(v2.x, 0, v2.y);
-    }
-    
-    private Vector3 ComputeCollisionMove(Vector3 desiredMove)
-    {
-        var output = desiredMove;
-        
-        // Radius of your character (adjust as needed)
-        float radius = CollisionMoveSphereCastRadius;
-        float castDistance = (CollisionMoveSphereCastDistance * GetRaycastTimeModifier()) - (radius * 0.45f);
-        
-        Vector3 position = transform.position + Vector3.up * (YVelocity > FallingCollisionMoveSphereCastHeightYVelocityThreshhold
-            ? GroundCollisionMoveSphereCastHeight
-            : FallingCollisionMoveSphereCastHeight);
-        Vector3 direction = output.normalized;
-
-        // SphereCast to account for player volume
-        if (Physics.SphereCast(position, radius, direction, out RaycastHit hit, castDistance, ~0, QueryTriggerInteraction.Ignore))
-        {
-            
-            // First collision: slide along the surface
-            Vector3 firstNormal = hit.normal;
-            output = Vector3.ProjectOnPlane(output, Vector3.ProjectOnPlane(firstNormal, Vector3.up));
-
-
-            // Cast again in the new direction to handle corner (second surface)
-            if (Physics.SphereCast(position, radius, output.normalized, out RaycastHit secondHit, output.magnitude))
-            {
-                Vector3 secondNormal = secondHit.normal;
-
-                // Slide again
-                output = Vector3.ProjectOnPlane(output, Vector3.ProjectOnPlane(secondNormal, Vector3.up));
-                
-                if (output.magnitude < 0.01f)
-                {
-                    output = Vector3.zero;
-                }
-
-            }
-            
-            
-        }
-        
-        return output;
-    }
-    
-    private float ComputeMomentumWeight()
-    {
-        return Mathf.InverseLerp(0, MaxMomentum, _momentum);
-    }
-
-    private Vector3 ComputeDesiredMove()
-    {
-        var value = Mathf.Lerp(0f, MaximumMomentumSpeedMod, ComputeMomentumWeight());
-        return transform.forward.normalized * (MoveSpeed * value * Time.deltaTime);
-    }
-
-    private void SetAnimatorMomentum()
-    {
-        Animator.SetFloat("Momentum", ComputeMomentumWeight());
-    }
-
-    private void HandleCollisionMove(float modifier = 1f)
-    {
-        var desiredMove = ComputeDesiredMove();
-        var collisionMove = ComputeCollisionMove(desiredMove);
-        transform.position += collisionMove * modifier;
-        
-        if (Machine.IsInState(PlayerFsmState.Wallsquat)) return;
-        if (Machine.IsInState(PlayerFsmState.SlowVaultHang)) return;
-        if (Machine.IsInState(PlayerFsmState.Wallrun)) return;
-        if (Machine.IsInState(GravityFsmState.Aerial) && YVelocity > WallsquatMinimumYVelocity - 1f) return;
-        
-        var collisionRatio = (desiredMove.magnitude + 1f) / (collisionMove.magnitude + 1f);
-        _momentum = Mathf.Max(0, _momentum - (MomentumLossRate * Time.deltaTime * (collisionRatio - 1f) * CollisionMomentumLossRate));
-    }
-
-    public void InvokeBoost(bool jump, float momentumWeight)
-    {
-        if (jump) Machine.Fire(PlayerFsmTrigger.Jump);
-        _momentum = MaxMomentum * momentumWeight;
-    }
-    
-    public void InvokeCheckpoint(Vector3 position, Quaternion rotation)
-    {
-        _checkpointVector3 = position;
-        _checkpointQuaternion = rotation;
-    }
-    
-    public static Vector3 MirrorInputForward(Vector3 input, Vector3 forward)
-    {
-        if (input == Vector3.zero)
-            return Vector3.zero;
-
-        // Project both vectors onto the XZ plane (ignore vertical component)
-        Vector3 inputFlat = new Vector3(input.x, 0f, input.z).normalized;
-        Vector3 forwardFlat = new Vector3(forward.x, 0f, forward.z).normalized;
-
-        float dot = Vector3.Dot(inputFlat, forwardFlat);
-
-        if (dot >= 0f)
-        {
-            // Input is within 90 degrees of Forward — return as-is
-            return input;
-        }
-        else
-        {
-            // Mirror the input vector across the forward's perpendicular plane
-            // First, get the right vector (90° rotation from forward)
-            Vector3 right = Vector3.Cross(Vector3.up, forwardFlat).normalized;
-
-            // Project input onto the forward-right basis
-            float f = Vector3.Dot(inputFlat, forwardFlat);
-            float r = Vector3.Dot(inputFlat, right);
-
-            // Mirror the forward component (flip sign of f)
-            float mirroredF = -f;
-
-            // Reconstruct the mirrored vector
-            Vector3 mirrored = (mirroredF * forwardFlat) + (r * right);
-
-            // Scale by original input magnitude (preserve intensity)
-            return mirrored.normalized * input.magnitude;
-        }
-    }
-
-    private void OnDrawGizmos()
-    {
-        Gizmos.color = Color.green;
-        Gizmos.DrawSphere(_currentLedgePosition, 0.25f);
-    }
-    
-    private bool CanGrapple(TriggerParams? triggerParams)
-    {
-        return PlayerWeaponFsm.Singleton.Machine.IsInState(PlayerWeaponFsm.PlayerWeaponFsmState.ImpaleStuck);
-    }
-    
-    private bool CanImpale(TriggerParams? triggerParams)
-    {
-        return PlayerWeaponFsm.Singleton.Machine.IsInState(PlayerWeaponFsm.PlayerWeaponFsmState.Idle);
-    }
-
-    private void OnContactHitboxCollide()
-    {
-        Machine.Fire(PlayerFsmTrigger.ContactHitboxTrigger);
+        return false;
     }
 
     private void OnEnable()
